requirements.txt
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.3.0
pypdf==3.17.0
